This code is now complete
It carries out NMF topic modelling
and then dumps a list of top words for each topic and a topic label
for each document to a feather binary.

Do note that ties are just ignored, it picks the first. I don't
think this is a big deal but I haven't thought about it much