Working commit of dashboard
This commit works, it's deployed in the testing area.
It draws a treemap, with the words per topic on the rectangle.
And you can click for comment display

Obviously it needs porting into a realistic example where you see
everything from the area in which you are interested, not just
10 random ones